refactor(CreatePayment): extract usePaymentAmountsEditor hook (SDK-1124)
Lifts line-item state, totals, edit-modal form and handlers into
usePaymentAmountsEditor in Contractor/Payments/shared/. Hook is
parameterized by allowedPaymentMethods; Root passes ['Check', 'Direct
Deposit']. Public API and all six events unchanged.
